requirements.txt
Flask==3.1.1
gunicorn==21.2.0
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
python-binance==1.0.29
pandas==2.3.1
plotly==6.2.0
ta==0.11.0
sqlalchemy==2.0.41 